+++ AgroAgil/app/src/main/java/com/example/agroagil/core/models/Product.kt
@@ -4,9 +4,9 @@ data class Product (
     val name: String = "",
     var amount: Float = 0f,
     var units: String = "",
-    val price: Double = 0.0
+    val price: Float = 0f
 ){
-    constructor(name: String= "", amount: Int = 0, units: String= "", price: Double=0.0):
+    constructor(name: String= "", amount: Int = 0, units: String= "", price: Float=0f):
             this(name, amount.toFloat(), units, price)
 
 }
